day học ngày 15/9/2018

+++ estate-platform-core/src/main/java/com/estate/repository/custom/BuildingRepositoryCustom.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface BuildingRepositoryCustom {
     List<BuildingEntity> findAll(BuildingBuilder buildingBuilder, Pageable pageable);
-    Integer getTotalItems(BuildingBuilder buildingBuilder);
+    Long getTotalItems(BuildingBuilder buildingBuilder);
     List<BuildingEntity> findByStaffs_id(BuildingBuilder buildingBuilder, Pageable pageable,long userId);
     Integer getTotalItemsByStaffs_id(BuildingBuilder buildingBuilder, long userId);
 }
